% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{R98Data}
\alias{R98Data}
\title{Title for R98Data}
\format{
A data frame with 15 variables:
\describe{
\item{Address}{Description of var1}
\item{MahMorajeh}{Description of var2}
\item{Fasl}{Description of var3}
\item{weight}{Description of var4}
\item{NoeKhn}{Description of var5}
\item{Takmil}{Description of var6}
\item{TakmilDescA}{Description of var7}
\item{TakmilDescB}{Description of var8}
\item{TakmilDescC}{Description of var9}
\item{Jaygozin}{Description of var10}
\item{JaygozinDescA}{Description of var11}
\item{JaygozinDescB}{Description of var12}
\item{JaygozinDescC}{Description of var13}
\item{BlkAbdJaygozin}{Description of var14}
\item{RadifJaygozin}{Description of var15}
}
}
\usage{
R98Data
}
\description{
Title for R98Data
}
\keyword{datasets}
