% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{R98P3S05}
\alias{R98P3S05}
\title{Table R98P3S05 \cr
Household furniture & appliance expenditure of rural household in 1398 \cr
هزينه هاي  مبلمان و لوازم خانگي خانوار روستایی سال 1398}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
R98P3S05
}
\description{
{R for Rural} \cr
{98 for 1398} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
