% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{U98P3S09}
\alias{U98P3S09}
\title{Table U98P3S09 \cr
Household recreation & leisure expenditure of urban household in 1398 \cr
هزينه هاي  مربوط به خدمات فرهنگي و تفريحات خانوار شهری سال 1398}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
U98P3S09
}
\description{
{U for Urban} \cr
{98 for 1398} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
